mqtt_proxy/requirements.txt
paho-mqtt
kafka-python